requirements.txt
absl-py==0.12.0
antlr4-python3-runtime==4.8
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
git+https://github.com/facebookresearch/detectron2.git@v0.2#egg=detectron2
future==0.18.2
fvcore==0.1.2.post20210128
google-auth==1.28.0
google-auth-oauthlib==0.4.3
grpcio==1.36.1
idna==2.10
imageio==2.9.0
importlib-metadata==3.8.0
injector==0.18.4
iopath==0.1.8
kiwisolver==1.3.1
layoutparser==0.1.3
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.0
mock==4.0.3
networkx==2.5
numpy==1.20.1
oauthlib==3.1.0
omegaconf==2.1.0.dev22
opencv-python==4.5.1.48
pandas==1.2.3
Pillow==8.1.2
pkg-resources==0.0.0
portalocker==2.2.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.1
pydot==1.4.2
pyenchant==3.2.0
pyparsing==2.4.7
PyQt5==5.12.2
PyQt5-sip==12.8.1
pytesseract==0.3.7
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
reportlab==3.5.56
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.6.2
six==1.15.0
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tifffile==2021.3.17
torch==1.8.1
torchvision==0.9.1
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
yacs==0.1.8
zipp==3.4.1
